+++ src/plugins/index.ts
@@ -5,6 +5,7 @@ const options = JSON.parse(`<%= JSON.stringify(options) %>`)
 const {
   registerUrl,
   loginUrl,
+  refreshUrl,
   userUrl,
   editUrl,
   expires,
@@ -27,7 +28,7 @@ const myPlugin: Plugin = (ctx, inject) => {
     },
     login({ username, password }: any): Promise<unknown> {
       return new Promise((resolve, reject) => {
-        ctx.$axios
+        ctx.app.$axios
           .$post(loginUrl, { username, password })
           .then((response) => {
             if (response[tokenProperty]) {
@@ -39,7 +40,11 @@ const myPlugin: Plugin = (ctx, inject) => {
               )
               storage.setUniversal('expiredAt', expired)
               storage.setUniversal('tokenType', tokenType)
-
+              ctx.app.$axios.setToken(
+                storage.state.token,
+                storage.state.tokenType
+              )
+              this.refreshToken()
               resolve(this.fetchUserData())
             } else {
               resolve(response)
@@ -52,8 +57,7 @@ const myPlugin: Plugin = (ctx, inject) => {
     },
     editInfo(data: object): Promise<Object | void> {
       return new Promise((resolve, reject) => {
-        ctx.$axios.setToken(storage.state.token, storage.state.tokenType)
-        ctx.$axios
+        ctx.app.$axios
           .$put(editUrl, data)
           .then((response) => {
             resolve(response)
@@ -65,8 +69,7 @@ const myPlugin: Plugin = (ctx, inject) => {
     },
     fetchUserData(): Promise<unknown> {
       return new Promise((resolve, reject) => {
-        ctx.$axios.setToken(storage.state.token, storage.state.tokenType)
-        ctx.$axios
+        ctx.app.$axios
           .$get(userUrl)
           .then((response) => {
             if (response) {
@@ -101,6 +104,34 @@ const myPlugin: Plugin = (ctx, inject) => {
     },
     getUserData(): unknown {
       return storage.state.user
+    },
+    getToken() {
+      return { type: storage.state.tokenType, token: storage.state.token }
+    },
+    refreshToken() {
+      return new Promise((resolve, reject) => {
+        const refreshToken = storage.getUniversal('refreshToken')
+        ctx.app.$axios
+          .$post(refreshUrl, { refresh_token: refreshToken })
+          .then((response) => {
+            if (response) {
+              const expired = new Date().getTime() + expires * 1000
+              storage.setUniversal('token', response[tokenProperty])
+              storage.setUniversal(
+                'refreshToken',
+                response[refreshTokenProperty]
+              )
+              storage.setUniversal('expiredAt', expired)
+              storage.setUniversal('tokenType', tokenType)
+              resolve({ success: true })
+            } else {
+              resolve(response)
+            }
+          })
+          .catch((error) => {
+            reject(error)
+          })
+      })
     },
   }
   inject('authCustom', allFn)